Let's just not use webpack alias as it acts up - let's just rename to @hedvig-ui instead :~)
