+++ .storybook/preview.ts
@@ -1,4 +1,4 @@
 import { addDecorator } from '@storybook/react'
-import { withTheme } from '../shared/hedvig-ui/storybook-utils'
+import { withTheme } from '../shared/@hedvig-ui/storybook-utils'
 
 addDecorator(withTheme)
