+++ shared/@hedvig-ui/DateTimePicker/date-time-picker.tsx
@@ -2,7 +2,6 @@ import styled from '@emotion/styled'
 import enGB from 'date-fns/locale/en-GB'
 import React from 'react'
 import DatePicker, { registerLocale } from 'react-datepicker'
-import 'react-datepicker/dist/react-datepicker.css'
 import { Input } from 'semantic-ui-react'
 
 registerLocale('enGB', enGB)
