--- shared/hedvig-ui/index.ts
@@ -1,148 +1,0 @@
-import { Badge, BadgeProps } from './Badge/badge'
-import {
-  Button,
-  buttonColorMap,
-  ButtonLink,
-  ButtonProps,
-  ButtonsGroup,
-  buttonSizeMap,
-} from './Button/button'
-import {
-  Card,
-  CardContent,
-  CardLink,
-  CardsWrapper,
-  DangerCard,
-} from './Card/card'
-import { CasualList, CasualListItem } from './CasualList/casual-list'
-import { Checkbox } from './Checkbox/checkbox'
-import { Copyable } from './Copyable/copyable'
-import { DateTimePicker } from './DateTimePicker/date-time-picker'
-import {
-  Dropdown,
-  EnumDropdown,
-  getTextFromEnumValue,
-} from './Dropdown/dropdown'
-import {
-  Form,
-  FormDropdown,
-  FormInput,
-  FormTextArea,
-  SubmitButton,
-} from './Form/form'
-import {
-  InfoContainer,
-  InfoRow,
-  InfoSection,
-  InfoTag,
-  InfoTagStatus,
-  InfoText,
-} from './InfoRow/info-row'
-import { CustomInputProps, Input } from './Input/input'
-import { JsonSchemaForm } from './JsonSchemaForm/json-schema-form'
-import { List, ListItem } from './List/list'
-import { Loadable } from './Loadable/loadable'
-import { FlagOrbIndicator, OrbIndicator } from './OrbIndicator/orb-indicator'
-import { Popover } from './Popover/popover'
-import { RadioGroup } from './Radio/radio'
-import { SearchableDropdown } from './SearchableDropdown/searchable-dropdown'
-import { Spinner } from './Spinner/spinner'
-import { TextArea } from './TextArea/text-area'
-
-import { FadeIn, withFadeIn } from './animations/fade-in'
-import {
-  LoadingMessage,
-  StandaloneMessage,
-} from './animations/standalone-message'
-import { Spacing, spacingMap, SpacingSize } from './Spacing/spacing'
-import {
-  darkTheme,
-  darkUiTheme,
-  lightTheme,
-  lightUiTheme,
-  SemanticOverrides,
-} from './themes'
-import {
-  Bold,
-  Capitalized,
-  ErrorText,
-  FourthLevelHeadline,
-  Label,
-  MainHeadline,
-  Paragraph,
-  Placeholder,
-  SecondLevelHeadline,
-  Shadowed,
-  ThirdLevelHeadline,
-} from './Typography/typography'
-
-export {
-  Badge,
-  Button,
-  ButtonsGroup,
-  buttonSizeMap,
-  BadgeProps,
-  buttonColorMap,
-  ButtonLink,
-  ButtonProps,
-  Checkbox,
-  Card,
-  CardContent,
-  CardLink,
-  CardsWrapper,
-  CasualList,
-  CasualListItem,
-  DangerCard,
-  Dropdown,
-  EnumDropdown,
-  getTextFromEnumValue,
-  Copyable,
-  DateTimePicker,
-  Form,
-  SubmitButton,
-  FormDropdown,
-  FormInput,
-  FormTextArea,
-  InfoRow,
-  InfoContainer,
-  InfoSection,
-  InfoTag,
-  InfoTagStatus,
-  InfoText,
-  Input,
-  CustomInputProps,
-  List,
-  ListItem,
-  JsonSchemaForm,
-  Loadable,
-  Popover,
-  OrbIndicator,
-  FlagOrbIndicator,
-  RadioGroup,
-  SearchableDropdown,
-  Spinner,
-  Spacing,
-  SpacingSize,
-  spacingMap,
-  TextArea,
-  FourthLevelHeadline,
-  ThirdLevelHeadline,
-  MainHeadline,
-  Shadowed,
-  SecondLevelHeadline,
-  Placeholder,
-  Paragraph,
-  Label,
-  Bold,
-  Capitalized,
-  ErrorText,
-  StandaloneMessage,
-  LoadingMessage,
-  FadeIn,
-  withFadeIn,
-  darkTheme,
-  darkUiTheme,
-  lightTheme,
-  lightUiTheme,
-  SemanticOverrides,
-}
