+++ webpack/webpack.config.base.js
@@ -20,10 +20,8 @@ module.exports = ({
     modules: [
       path.resolve(context, 'node_modules'),
       path.resolve(context, 'src'),
+      path.resolve(context, 'shared'),
     ],
-    alias: {
-      '@hedvig-ui': path.resolve(context, 'shared/hedvig-ui'),
-    },
   },
   entry,
   module: {
